cleaned App.tsx and created an other component
