+++ react-phase-1/src/components/LoginForm.tsx
@@ -1,0 +1,55 @@
+import React, { useState } from 'react';
+import './CustomerTable.css'; // Reuse table styling for consistency
+
+interface LoginFormProps {
+  onLogin: (email: string, password: string) => void;
+}
+
+const LoginForm: React.FC<LoginFormProps> = ({ onLogin }) => {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    onLogin(email, password);
+  };
+
+  return (
+    <div className="card mt-4" style={{ maxWidth: 400, margin: '0 auto' }}>
+      <div className="card-header">
+        <h5 className="mb-0">Login</h5>
+      </div>
+      <div className="card-body">
+        <form onSubmit={handleSubmit} data-testid="login-form">
+          <div className="mb-3">
+            <label htmlFor="loginEmail" className="form-label">Email</label>
+            <input
+              type="email"
+              className="form-control"
+              id="loginEmail"
+              value={email}
+              onChange={e => setEmail(e.target.value)}
+              placeholder="Enter email"
+              required
+            />
+          </div>
+          <div className="mb-3">
+            <label htmlFor="loginPassword" className="form-label">Password</label>
+            <input
+              type="password"
+              className="form-control"
+              id="loginPassword"
+              value={password}
+              onChange={e => setPassword(e.target.value)}
+              placeholder="Enter password"
+              required
+            />
+          </div>
+          <button type="submit" className="btn btn-primary w-100">Login</button>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default LoginForm;
